move out token config
