+++ spaceward/src/features/modals/util.ts
@@ -27,10 +27,10 @@ import erc20Abi from "@/contracts/eip155/erc20Abi";
 import { getProvider, isSupportedNetwork } from "@/lib/eth";
 import type { BalanceEntry } from "../assets/types";
 import { getAbiItem } from "../assets/util";
-import { COSMOS_CHAINS } from "../assets/hooks";
 import { encodeSecp256k1Pubkey, makeSignDoc, StdSignDoc } from "@cosmjs/amino";
 import { Int53 } from "@cosmjs/math";
 import { SignMode } from "@wardenprotocol/wardenjs/codegen/cosmos/tx/signing/v1beta1/signing";
+import { COSMOS_CHAINS } from "@/config/tokens";
 
 function typedStartsWith<T extends string>(
 	prefix: T,
